Git Audit - Add scm providers (#297)
